requirements.txt
djangorestframework
psycopg2-binary
django-environ
django-hosts
gunicorn
django